Moved functions, replaced functions with variables
Organized the Add Task function in its own file and replaced repetitive declarations of Firestore functions
