+++ lib/homepage.dart
@@ -3,12 +3,11 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:provider/provider.dart';
-import 'package:smarttodo/authentication/services/authentication_service.dart';
-import 'package:smarttodo/constants.dart';
+import 'package:smarttodo/add_task.dart';
+import 'package:smarttodo/functions.dart';
+import 'package:smarttodo/shared/constants.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 
@@ -30,6 +29,7 @@ class _HomePageState extends State<HomePage> {
   final formKeySaveTask = GlobalKey<FormState>();
   late String taskDocID = '';
   bool isEditModeEnabled = false;
+
 
   @override
   void initState() {
@@ -57,7 +57,7 @@ class _HomePageState extends State<HomePage> {
           padding: const EdgeInsets.all(0),
           child: const Icon(FontAwesomeIcons.rightToBracket),
           onPressed: () {
-            logoutConfirmationPrompt();
+            logoutConfirmationPrompt(context);
           }
         ),
       ),
@@ -75,10 +75,7 @@ class _HomePageState extends State<HomePage> {
                 children: [
                   Expanded(
                     child: StreamBuilder(
-                      stream: FirebaseFirestore.instance
-                        .collection('users')
-                        .doc(myUID(context))
-                        .collection('myTasks')
+                      stream: myTasksDBCollection(context)
                         .where('isTaskCompleted', isEqualTo: false)
                         .orderBy('dueDate')
                         .limit(1)
@@ -104,30 +101,21 @@ class _HomePageState extends State<HomePage> {
 
 
                               deleteTask(){
-                                FirebaseFirestore.instance
-                                  .collection('users')
-                                  .doc(myUID(context))
-                                  .collection('myTasks')
+                                myTasksDBCollection(context)
                                   .doc(taskDocID)
                                   .delete();
                               }
         
                               markTaskAsCompleted(){
-                                FirebaseFirestore.instance
-                                  .collection('users')
-                                  .doc(myUID(context))
-                                  .collection('myTasks')
+                                myTasksDBCollection(context)
                                   .doc(taskDocID)
                                   .update({
-                                    'isTaskCompleted': true
+                                    'isTaskCompleted': true,
                                   });
                               }
 
                               skipTask(){
-                                FirebaseFirestore.instance
-                                .collection('users')
-                                .doc(myUID(context))
-                                .collection('myTasks')
+                                myTasksDBCollection(context)
                                 .doc(taskDocID)
                                 .update({
                                   'timesSkipped': timesSkipped + 1,
@@ -136,10 +124,7 @@ class _HomePageState extends State<HomePage> {
                               }
 
                               editTask(){
-                                  FirebaseFirestore.instance
-                                  .collection('users')
-                                  .doc(myUID(context))
-                                  .collection('myTasks')
+                                  myTasksDBCollection(context)
                                   .doc(taskDocID)
                                   .update({
                                     'taskTile': editTaskTitleController.text,
@@ -700,8 +685,7 @@ class _HomePageState extends State<HomePage> {
                   child: const Icon(FontAwesomeIcons.bars,
                   color: CupertinoColors.white,),
                   onPressed: (){
-                    _showTaskListPanel(context);
-                    // Navigator.push(context, CupertinoPageRoute(builder: (context) => ListOfTasks()));
+                    showTaskListPanel(context);
                   },
                 ),
               ),
@@ -1000,327 +984,8 @@ Future<void> _launchInBrowser(Uri url) async {
   }
 
 
-  _showTaskListPanel(BuildContext context) {
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      backgroundColor: Colors.transparent,
-      builder: (context) {
-        return GestureDetector(
-          onTap: () => Navigator.of(context).pop(),
-          child: Container(
-            color: const Color.fromRGBO(0, 0, 0, 0.001),
-            child: GestureDetector(
-              onTap: () {},
-              child: DraggableScrollableSheet(
-                initialChildSize: 0.72,
-                minChildSize: 0.2,
-                builder: (_, controller) {
-                  return Container(
-                    height: 600,
-                    decoration: const BoxDecoration(
-                      color: Color(0xFF28293d),
-                      borderRadius: BorderRadius.only(
-                        topLeft: Radius.circular(25.0),
-                        topRight: Radius.circular(25.0),
-                      ),
-                    ),
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        const SizedBox(height: 15,),
-                        const Text('My Tasks',
-                          style: TextStyle(
-                            fontSize: 24,
-                            color: CupertinoColors.white,
-                            fontWeight: FontWeight.bold
-                          ),
-                        ),
-                        SizedBox(
-                            height: MediaQuery.of(context).size.height * 0.55,
-                            child: Padding(
-                              padding: const EdgeInsets.only(top: 0, bottom: 0, left: 0, right: 8),
-                              child: Row(
-                                children: [
-                                  Expanded(
-                                    child: Padding(
-                                      padding: const EdgeInsets.all(8.0),
-                                      child: StreamBuilder(
-                                        stream: FirebaseFirestore.instance
-                                          .collection('users')
-                                          .doc(myUID(context))
-                                          .collection('myTasks')
-                                          // .where('isTaskCompleted', isEqualTo: false)
-                                          .orderBy('isTaskCompleted')
-                                          .snapshots(),
-                                        builder: (BuildContext context, AsyncSnapshot<QuerySnapshot> snapshot) {
-                                          if (snapshot.hasData && snapshot.data!.docs.isNotEmpty) {
-                                            return ListView.builder(
-                                              physics: const BouncingScrollPhysics(),
-                                              scrollDirection: Axis.vertical,
-                                              itemCount: snapshot.data?.docs.length,
-                                              itemBuilder: (BuildContext context, int index) {
-                                                QueryDocumentSnapshot<Object?>? user = snapshot.data?.docs[index];
-                          
-                                                var taskTile = user!.data().toString().contains('taskTile') ? user.get('taskTile') : '';
-                                                var taskDocID = user.data().toString().contains('taskDocID') ? user.get('taskDocID') : '';
-                                                var isTaskCompleted = user.data().toString().contains('isTaskCompleted') ? user.get('isTaskCompleted') : false;
-                  
-                                                return Column(
-                                                  children: [
-                                                    Slidable(
-                                                      key: const ValueKey(0),
-                                                      endActionPane:  ActionPane(
-                                                      motion: const StretchMotion(),
-                                                      children: [
-                                                        SlidableAction(
-                                                          autoClose: true,
-                                                          onPressed: (BuildContext context) async {
-                                                            FirebaseFirestore.instance
-                                                            .collection('users')
-                                                            .doc(myUID(context))
-                                                            .collection('myTasks')
-                                                            .doc(taskDocID)
-                                                            .delete();
-                                                          },
-                                                          backgroundColor: CupertinoColors.systemRed,
-                                                          foregroundColor: Colors.white,
-                                                          icon: FontAwesomeIcons.trashCan,
-                                                          borderRadius: const BorderRadius.all(Radius.circular(5.0)),
-                                                        ),
-                                                      ],
-                                                    ),
-                                                      child: Builder(
-                                                        builder: (context) {
-                                                          if (isTaskCompleted == false) {
-                                                            return Row(
-                                                            children: [
-                                                              Expanded(
-                                                                child: GestureDetector(
-                                                              onTap: () {
-                                                                FirebaseFirestore.instance
-                                                                  .collection('users')
-                                                                  .doc(myUID(context))
-                                                                  .collection('myTasks')
-                                                                  .doc(taskDocID)
-                                                                  .update({
-                                                                    'isTaskCompleted': true,
-                                                                  });
-                                                              },
-                                                              child: Padding(
-                                                                padding: const EdgeInsets.all(5),
-                                                                child: Row(
-                                                                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                                                  children: [
-                                                                    Padding(
-                                                                      padding: const EdgeInsets.only(top: 0, bottom: 0, left: 5, right: 5),
-                                                                      child: Column(
-                                                                        children: [
-                                                                          SizedBox(
-                                                                            width: 30,
-                                                                            height: 30,
-                                                                            child: Row(
-                                                                              mainAxisAlignment: MainAxisAlignment.center,
-                                                                              children: const [
-                                                                                Icon(FontAwesomeIcons.circle,
-                                                                                  size: 26,
-                                                                                  
-                                                                                )
-                                                                              ],
-                                                                            ),
-                                                                          ),
-                                                                        ],
-                                                                      ),
-                                                                    ),
-                                                                    Expanded(
-                                                                      child: GestureDetector(
-                                                                        onTap: () {
-                                                                          FirebaseFirestore.instance
-                                                                          .collection('users')
-                                                                          .doc(myUID(context))
-                                                                          .collection('myTasks')
-                                                                          .doc(taskDocID)
-                                                                          .update({
-                                                                            'isTaskCompleted': true,
-                                                                          });
-                                                                        },
-                                                                        child: Column(
-                                                                          crossAxisAlignment: CrossAxisAlignment.start,
-                                                                          children: [
-                                                                            Text(taskTile,
-                                                                              style: const TextStyle(
-                                                                                color: CupertinoColors.white,
-                                                                                fontWeight: FontWeight.bold,
-                                                                              ),
-                                                                              maxLines: 1,
-                                                                              overflow: TextOverflow.ellipsis,
-                                                                            ),
-                                                                          ],
-                                                                        ),
-                                                                      ),
-                                                                    ),
-                                                                  ],
-                                                                ),
-                                                              ),
-                                                                ),
-                                                              ),
-                                                            ],
-                                                          );
-                                                            
-                                                          } else {
-                                                            return Row(
-                                                              children: [
-                                                                Expanded(
-                                                                  child: GestureDetector(
-                                                                onTap: () {
-                                                                FirebaseFirestore.instance
-                                                                  .collection('users')
-                                                                  .doc(myUID(context))
-                                                                  .collection('myTasks')
-                                                                  .doc(taskDocID)
-                                                                  .update({
-                                                                    'isTaskCompleted': false,
-                                                                  });
-                                                                },
-                                                                child: Padding(
-                                                                  padding: const EdgeInsets.all(5),
-                                                                  child: Row(
-                                                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                                                    children: [
-                                                                      Padding(
-                                                                        padding: const EdgeInsets.only(top: 0, bottom: 0, left: 5, right: 5),
-                                                                        child: Column(
-                                                                          children: [
-                                                                            SizedBox(
-                                                                              width: 30,
-                                                                              height: 30,
-                                                                              child: Row(
-                                                                                mainAxisAlignment: MainAxisAlignment.center,
-                                                                                children: const [
-                                                                                  Icon(FontAwesomeIcons.solidCircleCheck,
-                                                                                    color: CupertinoColors.activeGreen,
-                                                                                    size: 26,
-                                                                                  )
-                                                                                ],
-                                                                              ),
-                                                                            ),
-                                                                          ],
-                                                                        ),
-                                                                      ),
-                                                                      Expanded(
-                                                                        child: GestureDetector(
-                                                                          onTap: () {
-                                                                            // do something when clicking on it
-                                                                    
-                                                                          },
-                                                                          child: Column(
-                                                                            crossAxisAlignment: CrossAxisAlignment.start,
-                                                                            children: [
-                                                                              Text(taskTile,
-                                                                                style: const TextStyle(
-                                                                                  decoration: TextDecoration.lineThrough,
-                                                                                  fontWeight: FontWeight.bold,
-                                                                                  color: CupertinoColors.white,
-                                                                                ),
-                                                                                maxLines: 1,
-                                                                                overflow: TextOverflow.ellipsis,
-                                                                              ),
-                                                                            ],
-                                                                          ),
-                                                                        ),
-                                                                      ),
-                                                                    ],
-                                                                  ),
-                                                                ),
-                                                                  ),
-                                                                ),
-                                                              ],
-                                                            );
-                                                          }
-                                                        }
-                                                      ),
-                                                    ),
-                                                    const Divider()
-                                                  ],
-                                                );
-                                              }
-                                            );
-                                          } if (snapshot.hasData && snapshot.data!.docs.isEmpty) {
-                                            // if the stream of data is empty
-                                            return const Center(
-                                              child: Text('no tasks to show'),
-                                            );
-                                          } else {
-                                            // Still loading
-                                            return loadingWidget;
-                                          }
-                                        },
-                                      ),
-                                    ),
-                                  ),
-                                ]
-                              ),
-                            ),
-                          ),
-
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            Expanded(
-                              child: Padding(
-                                padding: const EdgeInsets.all(8.0),
-                                child: Container(
-                                  decoration: BoxDecoration(
-                                    color: CupertinoColors.activeGreen,
-                                    borderRadius: const BorderRadius.all(
-                                      Radius.circular(35.0),
-                                    ),
-                                    boxShadow: [
-                                      BoxShadow(
-                                        color: CupertinoColors.black.withOpacity(0.1),
-                                        spreadRadius: 4,
-                                        blurRadius: 10,
-                                        offset: const Offset(0, 3),
-                                      )
-                                    ]
-                                  ),
-                                  child: CupertinoButton(
-                                    // padding: EdgeInsets.all(8),
-                                    child: const Text('Remove Completed',
-                                      style: TextStyle(
-                                        color: CupertinoColors.white,
-                                        fontWeight: FontWeight.bold
-                                      ),
-                                    ), 
-                                    onPressed: () {
-                                      deleteCompletedTasks();
-                                    }
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                        const SizedBox(
-                          height: 25,
-                        )
-                      ],
-                    ),
-                  );
-                },
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-
   saveTaskToFirestore(){
-    FirebaseFirestore.instance
-    .collection('users')
-    .doc(myUID(context))
-    .collection('myTasks')
+    myTasksDBCollection(context)
     .doc(taskDocID)
     .set({
       'taskTile': taskTitleController.text,
@@ -1332,63 +997,6 @@ Future<void> _launchInBrowser(Uri url) async {
       'timesSkipped': 1,
     });
   }
-
-
-  deleteCompletedTasks() async {
-    CollectionReference ref = 
-    FirebaseFirestore.instance
-      .collection('users')
-      .doc(myUID(context))
-      .collection('myTasks');
-
-    QuerySnapshot eventsQuery = await ref.where('isTaskCompleted', isEqualTo: true).get();
-
-    eventsQuery.docs.forEach((value) {
-    value.reference.delete();
-    });
-  }
-
-  logoutConfirmationPrompt(){
-    return showCupertinoDialog(
-      context: context,
-      builder: (_) {
-        return CupertinoAlertDialog(
-          title: const Text('Are you sure you want to logout?'),
-          actions: [
-            CupertinoDialogAction(
-              child: const Text('Yes, logout',
-                style: TextStyle(
-                  color: CupertinoColors.systemRed,
-                  fontWeight: FontWeight.bold,
-                  fontFamily: 'Nunito'
-                ),
-              ),
-              onPressed: () async {
-                await context.read<AuthenticationService>().signOut();
-                Navigator.of(context, rootNavigator: true).pop();
-              },
-            ),
-            CupertinoDialogAction(
-              child: const Text('Cancel',
-                style: TextStyle(
-                  color: CupertinoColors.white,
-                  fontWeight: FontWeight.bold,
-                  fontFamily: 'Nunito'
-                ),
-              ),
-              onPressed: () async {
-                Navigator.of(context, rootNavigator: true).pop();
-              },
-            ),
-          ]
-        );
-        
-      },
-      barrierDismissible: true,
-    );
-  }
-
-
 
 }
